Random mod logic
By some miracle I figured out how to randomly select a mod from a pool of mods, roll the tier of that mod, and roll the value in that tier. Mod and tier weights are also functioning so I can control how likely each mod and tier in that mod is to show up. Next goal is to support hybrid mods and to remove certain mods from available (already rolled, or prefix/suffix already full).