hybrid mods and no duplicate
I updated the add mod button so that hybrid mods are supported.  I also added some logic to prevent the same mod from being rolled once it's  already rolled (of course the same stat can still roll twice, once from the normal mod and once from hybrid, but this is intended). I have not worked out how to show the sum (if the same stat rolls on two mods) but I'm going to put that on the back burner right now as  it's not crucial to the functionality of the app and I have several other features I want to add first. Next up is capping suffix's and prefix's at 3 each and then I'll probably work on styling so it's not so ugly.