Fixed life regen affix display
The function that picks a random value in a tier range uses math.floor() so it doesn't play nice with decimal values, which Life Regen is. I have roll in the 100s and then just alter the display to divide by 100 before it shows so it looks normal. I had accidentally referenced the mod before the division outside of a for loop, which meant that if it rolled Life Regen early each new mod would divide it by 100 again, so it was showing some absurd numbers. Should be fixed now.