Annulment orb, item headers
I added the annulment orb as I felt it was important to have for normal crafting simulation. I had forgotten about divine orb so I'll probably add it next as I don't think the logic behind it is too difficult. I also have item headers in that display based on rarity so it's easier to tell which rarity the item is currently on. This helps because annulments can create some odd items (like 1 or even 0 affix rare items, haha). The images I used are low quality but they'll work for now until I take the time to get some better versions. At this point, I'm going to try to host this webapp on github so that a few people can try it out and inevitebly find a ton of bugs for me to fix.