Divine, currency counter, options
Added the divine orb and it's functionality. Also added a currency counter! It keeps track of every orb you use and can be reset if you choose with a reset button. I've also added a configuration state to handle different menus showing and certain options toggled. There is an option button now which changes the dynamic display (where the currency counter resides) to an options menu but I haven't built the options menu yet. Next up is to build an options  menu which will at first just consist of a single toggle to show tier  and suffix/prefix. All future options that the user can control will go here.