Affix sorting, dynamic name
I added some logic to sort the affixs so that they are listed predictably. I couldn't really find a rhyme or reason for how GGG sorts their affixes, so I just looked at a bunch of rare items and figured out the order roughly. I'll fix mistakes in this order as I come across them but it looks a lot better. I also added logic for dynamic names! This means that magic items will have a name determined by the prefix and suffix that they have. This should help those that alt spam while watching the name for a clue as to when they hit a T1 affix that they want. Rare item name generation also works. Just like in the game, the prefix is chosen randomly from a large list and the suffix is chosen from a smaller list tied to the base item type. I believe only alchs/regals/chaos change the name of a rare but if this is wrong, it's easy to add the name changing functionality to other item crafts.