Finished MetaMods
Finally figured out all of the logic behind Prefixes and Suffixes cannot be changed, and how they react with the other orbs. Crafted stat is now counted towards the 6 affixes like it should. Annulment obeys the cannot be changed metamods if they are present. Chaos orb also obeys the metamods, so it only rerolls mods not protected, which includes the metamod itself. Current system isn't really going to work for "can have multiple crafted mods" but I don't think I care about adding that anyways. It's not really needed for a proper simulation in my opinion.