Styled and fixed a few bugs
Tried to make the styling a little better. Added some margin to a few elements that were right on the edge, centered a few other elements. Changed the way the background images for the orbs scaled so that the bottom won't be cut off anymore. Also added a cost to the remove craft master mod.